Organize packets into packages (play phase)
