+++ StaffMod/src/main/kotlin/opekope2/avm_staff/internal/client/StaffModClient.kt
@@ -16,8 +16,8 @@ import net.minecraft.client.render.model.BakedModel
 import net.minecraft.client.util.InputUtil
 import opekope2.avm_staff.internal.StaffMod.MOD_ID
 import opekope2.avm_staff.internal.model.StaffItemModel
-import opekope2.avm_staff.internal.packet.c2s.AddItemToStaffC2SPacket
-import opekope2.avm_staff.internal.packet.c2s.RemoveItemFromStaffC2SPacket
+import opekope2.avm_staff.internal.packet.c2s.play.AddItemToStaffC2SPacket
+import opekope2.avm_staff.internal.packet.c2s.play.RemoveItemFromStaffC2SPacket
 import opekope2.avm_staff.util.isItemInStaff
 import org.lwjgl.glfw.GLFW
